Cleanup template/input
